pyproject.toml
[project]
name = "azure-102"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-textanalytics>=5.3.0",
    "azure-ai-translation-text>=1.0.1",
    "azure-ai-vision-imageanalysis>=1.0.0",
    "azure-cognitiveservices-speech>=1.47.0",
    "openai>=2.14.0",
    "python-dotenv>=1.0.0",
]
